Feature/update before 6

+++ work/services/before_install_services.py
@@ -3,6 +3,7 @@ from django.shortcuts import render, redirect
 from system_manager.models import InstallLocate
 from work.forms.before_install_form import BeforeInstallCheckListForm
 from django.contrib import messages
+from django.db.models import Q
 
 from work.models import (
     BeforeInspectionItem,
@@ -126,6 +127,8 @@ def update_before_checklist_service(request, type, pk):
     if request.method == "POST":
         form = BeforeInstallCheckListForm(request.POST, instance=instance)
         if form.is_valid():
+
+            # 기본 저장 영역
             before_checklist = form.save(commit=False)
             before_checklist.equipment = type
             if "locate" in request.POST.keys():
@@ -135,13 +138,26 @@ def update_before_checklist_service(request, type, pk):
             pk_checklist = request.POST.getlist("checklist-pk")
             image_keys = request.FILES.keys()
             before_checklist.save()
+
+            # 삭제 영역
+            before_list = before_checklist.before_inspection_result.all()
+            for before in before_list:
+                if str(before.before_inspection_item_id.pk) not in pk_checklist:
+                    before.delete()
+
+            # 수정 영역
             for pk_item in pk_checklist:
                 before_result_item = BeforeInspectionResult.objects.filter(
                     before_install_checklist_id=before_checklist,
                     before_inspection_item_id=BeforeInspectionItem.objects.get(
                         pk=pk_item
                     ),
-                )[0]
+                )
+                delete_before = False
+                # 수정된 영역이면 해당 구문이 실행, 새로 생성된 영역은 패스
+                if before_result_item:
+                    before_result_item = before_result_item[0]
+                    delete_before = True
                 result_item = BeforeInspectionResult(
                     result=request.POST[pk_item],
                     before_install_checklist_id=before_checklist,
@@ -165,7 +181,8 @@ def update_before_checklist_service(request, type, pk):
                     images = request.FILES.getlist(f"{pk_item}-images[]")
                     for img in images:
                         result_item.measures.create(img=img)
-                before_result_item.delete()
+                if delete_before:
+                    before_result_item.delete()
             messages.success(request, "저장이 완료되었습니다.")
             return redirect("work:update_before_install_checklist", type, pk)
     else:
